+++ lib/resty/lmdb/cdefs.lua
